+++ mod-record-specifications-server/src/main/java/org/folio/rspec/integration/kafka/SpecificationPartialChangedEventProducer.java
@@ -2,6 +2,8 @@ package org.folio.rspec.integration.kafka;
 
 import java.util.UUID;
 import org.folio.rspec.domain.dto.SpecificationUpdatedEvent;
+import org.folio.rspec.domain.repository.SpecificationRepository;
+import org.folio.rspec.exception.ResourceNotFoundException;
 import org.folio.spring.FolioExecutionContext;
 import org.springframework.context.annotation.Primary;
 import org.springframework.kafka.core.KafkaTemplate;
@@ -11,10 +13,13 @@ import org.springframework.stereotype.Component;
 @Component
 public class SpecificationPartialChangedEventProducer extends EventProducer<UUID, SpecificationUpdatedEvent> {
 
-  public SpecificationPartialChangedEventProducer(
-    KafkaTemplate<String, SpecificationUpdatedEvent> template,
-    FolioExecutionContext context) {
+  private final SpecificationRepository specificationRepository;
+
+  public SpecificationPartialChangedEventProducer(KafkaTemplate<String, SpecificationUpdatedEvent> template,
+                                                  SpecificationRepository specificationRepository,
+                                                  FolioExecutionContext context) {
     super(template, context);
+    this.specificationRepository = specificationRepository;
   }
 
   @Override
@@ -24,6 +29,18 @@ public class SpecificationPartialChangedEventProducer extends EventProducer<UUID
 
   @Override
   protected SpecificationUpdatedEvent buildEvent(UUID specificationId) {
-    return new SpecificationUpdatedEvent(specificationId, tenantId(), SpecificationUpdatedEvent.UpdateExtent.PARTIAL);
+    var specification = specificationRepository.findById(specificationId)
+      .orElseThrow(() -> ResourceNotFoundException.forSpecification(specificationId));
+    return new SpecificationUpdatedEvent(
+      specification.getId(),
+      tenantId(),
+      specification.getFamily(),
+      specification.getProfile(),
+      updateExtent()
+    );
+  }
+
+  protected SpecificationUpdatedEvent.UpdateExtent updateExtent() {
+    return SpecificationUpdatedEvent.UpdateExtent.PARTIAL;
   }
 }